pyproject.toml
[tool.poetry]
name = "TestCapitalCom"
version = "0.1.0"
description = ""
authors = ["Alexander Tomelo <qa.tomelo.an@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "7.4.0"
#webdriver-manager = "^3.8.6"
webdriver-manager = "^4.0.0"
selenium = "^4.10.0"
#selenium = "4.11.2"
pytest-html = "^3.2.0"
allure-pytest = "^2.12.0"
pytest-xdist = "^3.1.0"
#pyautogui = "^0.9.53"
psutil = "^5.9.4"
speedtest-cli = "^2.1.3"
#aiogram = "^2.24"
black = "23.7.0"
flake8 = "6.1.0"
#ruff
#memory-profiler

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
